requirements.txt
requests==2.32.3
bs4==0.0.2
python-slugify==8.0.4